backend/requirements.txt
fastapi==0.115.0
uvicorn==0.27.0
pydantic==2.5.3
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-multipart==0.0.22
python-dotenv==1.0.0
watchdog==3.0.0
markdown==3.5.1
pymdown-extensions==10.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bleach==6.1.0
